+++ src/CMakeLists.txt
@@ -219,6 +219,7 @@ add_library(gridcoin_util STATIC
     support/cleanse.cpp
     support/lockedpool.cpp
     sync.cpp
+    txmempool.cpp
     uint256.cpp
     util.cpp
     util/bip32.cpp
